Centralize logging code, de-static FunctionsEmulator
